Complete task 0

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -12,7 +12,7 @@ int main(void)
 	int i;
 
 	char my_string[] = "_putchar";
-	for (i = 0; i < strlen(my_string); i++)
+	for (i = 0; i < (int) strlen(my_string); i++)
 		_putchar(my_string[i]);
 	_putchar('\n');
 	return 0;
